manifest.py: create_initial: append rather than replace package lists
Presently, when the variables enumerated in Manifest.var_maps are
scanned for packages to add to the manifest (in
RpmManifest.create_initial() or OpkgManifest.create_initial()), the
existing package list for a particular pkg_type is overwritten. This
prevents us from adding new variables to var_maps that are the same
var_type as existing variables.

For instance, consider the variable PACKAGE_INSTALL which is of type
PKG_TYPE_MUST_INSTALL. Suppose that we added PACKAGE_INSTALL2 to
var_maps which also is of type PKG_TYPE_MUST_INSTALL. We would like to
populate the manifest both with PACKAGE_INSTALL and PACKAGE_INSTALL2,
but instead, only one or the other is used.

To make this work, create_initial() should append to package lists while
scanning var_maps instead of overwriting them.

Reviewed-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>
Reviewed-by: Haris Okanovic <haris.okanovic@ni.com>
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-Reviewboard-ID: 110700