dkms: Add new recipe for supporting dkms
Dkms is a tool used to compile out of tree modules directly on
target devices (sometimes you have sources or binaries which
need to be compiled directly on target, outside OE).
I used a simple module and a dkms config to test it, and dkms
compiled and installed the module on target.

For example:
.
├── dkms.conf
└── src
    ├── hello.c
    └── Makefile

hello.c:
	#include <linux/module.h>

	int init_module(void)
	{
		printk("Hello World!\n");
		return 0;
	}

	void cleanup_module(void)
	{
		printk("Goodbye Cruel World!\n");
	}

	MODULE_LICENSE("GPL");

Makefile:
	obj-m += hello.o

	KERNELDIR ?= /lib/modules/$(shell uname -r)/build

	PWD := $(shell pwd)

	install:
		$(MAKE) -C $(KERNELDIR) M=$(PWD) modules

	clean:
		rm -rf *.o *.ko *~ core .depend .*.cmd *.mod.c modules.order Module.symvers

dkms.conf:
	MAKE="make -C src/ KERNELDIR=/lib/modules/${kernelver}/build"
	CLEAN="make -C src/ clean"
	BUILT_MODULE_NAME=hello
	DEST_MODULE_LOCATION=/kernel/drivers/scsi
	BUILT_MODULE_LOCATION=src/
	PACKAGE_NAME=hello
	PACKAGE_VERSION=0.1
	AUTOINSTALL=yes
	REMAKE_INITRD=yes

All files should be in /usr/src/<package-name>-<package-version>,
in our case in /urs/src/hello-0.1.

Signed-off-by: Alexandru Vasiu <alexandru.vasiu@ni.com>