openembedded-core: Update umountnfs update-rc.d rule to stop.
The umountnfs script is currently configured to "start", and runs after
networking is stopped. This causes a ~3 minute timeout after networking is
shutdown when the system tries to connect to any mounted network shares.
Fix this by changing the script configuration to "stop" so that it can run
before networking is stopped.

Signed-off-by: Shruthi Ravichandran <shruthi.ravichandran@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>
Natinst-ReviewBoard-ID: 329863